+++ client/src/pages/Auth/SocialLogin.js
@@ -64,7 +64,8 @@ const SocialButton = styled.button`
 `;
 
 export default function SocialLogIn() {
-	const URI = 'https://wicked-husky-45.loca.lt/oauth2/authorization';
+	const URI =
+		'http://ec2-43-201-37-71.ap-northeast-2.compute.amazonaws.com:8080/oauth2/authorization';
 	const loginRequestHandler = (type) => {
 		window.location.assign((window.location.href = `${URI}/${type}`));
 	};
