+++ client/src/utils/axiosInstance.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const axiosInstance = axios.create({
-	baseURL: 'https://wicked-husky-45.loca.lt', // 서버 url
+	baseURL: 'http://ec2-43-201-37-71.ap-northeast-2.compute.amazonaws.com:8080', // 서버 url
 	timeout: 5000,
 });
